chart fix
