--- src/db/data.js
@@ -1,19 +1,0 @@
-const expensesData = [
-  { label: "Pork", value: 1000 },
-  { label: "Beef", value: 800 },
-  { label: "Chiken", value: 650 },
-  { label: "Fish", value: 600 },
-  { label: "Panini", value: 400 },
-  { label: "Coffee", value: 390 },
-  { label: "Spaghetti", value: 230 },
-  { label: "Chocolate", value: 200 },
-  { label: "Olives", value: 100 },
-  { label: "Greens", value: 50 },
-];
-
-const incomeData = [
-  { label: "Me", value: 25000 },
-  { label: "My wife", value: 20000 },
-];
-
-export { expensesData, incomeData };
