+++ src/helpers/descriptionTranslations.js
@@ -1,0 +1,9 @@
+const descriptionTranslation = {
+  ЗП: "Salary",
+  Коммуналка: "Utilities",
+  Кошти: "Money",
+  " Whiskey": "Whiskey",
+  "Birthday party": "Birthday party",
+};
+
+export default descriptionTranslation;
