+++ src/pages/Reports/Reports.jsx
@@ -9,6 +9,7 @@ import IncomeReports from "../../components/IncomeReports";
 import SwitchLeft from "../../components/Icons/SwitchLeft";
 import SwitchRight from "../../components/Icons/SwitchRight";
 import { getCurrentPeriod } from "../../helpers/getCurrentPeriod";
+import BarChart from "../../components/BarChart";
 
 const Reports = () => {
   const [dataType, setDataType] = useState("expenses");
@@ -52,6 +53,7 @@ const Reports = () => {
           {dataType === "expenses" && <ExpensesReports period={period} />}
           {dataType === "income" && <IncomeReports period={period} />}
         </div>
+        <BarChart period={period} dataType={dataType} />
         <Footer />
       </div>
     </div>
